Reorganize apis to use IDs
